refactor: split DriverEvent / WireMessage out of InboundConnectionDriver
The driver file was approaching the project's 500-line file-size guideline.
Hoists the dispatch-loop event union (DriverEvent) and the inbound-pump
message union (WireMessage) into a sister file so the driver itself can
focus on lifecycle orchestration.

Also tightens KDoc throughout the driver: the contract is the same, just
less wordy.
